Repo Creation
This will be a dungeon crawler esq game, made though simple picture
manipulation (ha ha, simple, if anything, I'll just make it crappy). It
will contain an inventory system item pickups, simple AI, health, and
several rooms. It is my sincere hope to create a "proper" game with
this, and not use simple trickery. I want picked up items, for instance,
to actually be picked up, and given to your character, not just walked
over and made invisible on the ground. Likewise for enemies. When they
die, I want their objects in memory to disappear. I am setting this as a
goal for myself, which hopefully I can succeed in.